chore: 패키지 구조 수정
패키지 구조를 레이어 기준에서 도메인 기준으로 변경
