--- src/main/java/com/souljit2/selleverything/config/RedisConfig.java
@@ -1,61 +1,0 @@
-package com.souljit2.selleverything.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.souljit2.selleverything.constants.CacheNames;
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
-
-@Configuration
-@EnableRedisHttpSession
-public class RedisConfig {
-
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    @Value("${spring.redis.host}")
-    private String redisHost;
-
-    @Value("${spring.redis.port}")
-    private int redisPort;
-
-    @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory(redisHost, redisPort);
-    }
-
-    @Bean
-    public RedisCacheManager redisCacheManager() {
-
-        RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration
-            .defaultCacheConfig()
-            .disableCachingNullValues()
-            .serializeValuesWith(
-                RedisSerializationContext.SerializationPair
-                    .fromSerializer(new GenericJackson2JsonRedisSerializer(objectMapper))
-            );
-
-        Map<String, RedisCacheConfiguration> redisCacheConfigurationMap = new HashMap<>();
-        redisCacheConfigurationMap
-            .put(CacheNames.POST, redisCacheConfiguration.entryTtl(Duration.ofMinutes(5)));
-
-        return RedisCacheManager.RedisCacheManagerBuilder
-            .fromConnectionFactory(redisConnectionFactory())
-            .withInitialCacheConfigurations(redisCacheConfigurationMap)
-            .cacheDefaults(redisCacheConfiguration)
-            .build();
-
-    }
-
-}
