--- src/main/java/com/souljit2/selleverything/constants/CacheNames.java
@@ -1,5 +1,0 @@
-package com.souljit2.selleverything.constants;
-
-public class CacheNames {
-    public static final String POST = "POST";
-}
