--- src/main/java/com/souljit2/selleverything/constants/DomainNames.java
@@ -1,6 +1,0 @@
-package com.souljit2.selleverything.constants;
-
-public class DomainNames {
-    public static String MEMBER = "member";
-    public static String POST = "post";
-}
