+++ src/main/java/com/souljit2/selleverything/domain/member/service/AuthService.java
@@ -1,0 +1,14 @@
+package com.souljit2.selleverything.domain.member.service;
+
+import com.souljit2.selleverything.domain.member.model.MemberDTO;
+import com.souljit2.selleverything.domain.member.model.SignInRequestDTO;
+
+public interface AuthService {
+
+    public void signUp(MemberDTO newMemberInfo);
+
+    public void signIn(SignInRequestDTO signInInfo);
+
+    public int getRequestMemberId();
+
+}
