+++ src/main/java/com/souljit2/selleverything/domain/member/service/MemberService.java
@@ -1,0 +1,11 @@
+package com.souljit2.selleverything.domain.member.service;
+
+import com.souljit2.selleverything.domain.member.model.MemberDTO;
+import com.souljit2.selleverything.domain.member.model.SignInRequestDTO;
+
+public interface MemberService {
+
+    public void insertMember(MemberDTO newMemberInfo);
+
+    public MemberDTO getMemberInfo(SignInRequestDTO signInInfo);
+}
