+++ src/main/java/com/souljit2/selleverything/domain/post/model/PostCategory.java
@@ -1,0 +1,11 @@
+package com.souljit2.selleverything.domain.post.model;
+
+public enum PostCategory {
+    CLOTHES,
+    INSTRUMENTS,
+    VEHICLES,
+    PHONES,
+    ACCESSORIES,
+    PROPS,
+    ETCS
+}
