+++ src/main/java/com/souljit2/selleverything/domain/post/service/PostService.java
@@ -1,0 +1,19 @@
+package com.souljit2.selleverything.domain.post.service;
+
+import com.souljit2.selleverything.domain.post.model.PostDTO;
+
+import java.util.List;
+import java.util.Map;
+
+public interface PostService {
+
+    public PostDTO getPostById(int id);
+
+    public List<PostDTO> getPostsByQueryString(Map<String, String> queryMap);
+
+    public void createPost(PostDTO newPost);
+
+    public void updatePostById(int id, PostDTO newPost);
+
+    public void deletePostById(int id);
+}
