--- src/main/java/com/souljit2/selleverything/exception/AuthenticationFailedException.java
@@ -1,10 +1,0 @@
-package com.souljit2.selleverything.exception;
-
-
-public class AuthenticationFailedException extends RuntimeException {
-
-    public AuthenticationFailedException(String msg) {
-        super(msg);
-    }
-
-}
