+++ src/main/java/com/souljit2/selleverything/global/config/WebMvcConfig.java
@@ -1,0 +1,20 @@
+package com.souljit2.selleverything.global.config;
+
+import com.souljit2.selleverything.global.interceptors.SessionCheckInterceptor;
+import lombok.AllArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@AllArgsConstructor
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    SessionCheckInterceptor sessionCheckInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(sessionCheckInterceptor)
+            .excludePathPatterns("/apis/auth/**");
+    }
+}
