+++ src/main/java/com/souljit2/selleverything/global/constants/CacheNames.java
@@ -1,0 +1,5 @@
+package com.souljit2.selleverything.global.constants;
+
+public class CacheNames {
+    public static final String POST = "POST";
+}
