+++ src/main/java/com/souljit2/selleverything/global/constants/DomainNames.java
@@ -1,0 +1,6 @@
+package com.souljit2.selleverything.global.constants;
+
+public class DomainNames {
+    public static String MEMBER = "member";
+    public static String POST = "post";
+}
