+++ src/main/java/com/souljit2/selleverything/global/exception/AuthenticationFailedException.java
@@ -1,0 +1,10 @@
+package com.souljit2.selleverything.global.exception;
+
+
+public class AuthenticationFailedException extends RuntimeException {
+
+    public AuthenticationFailedException(String msg) {
+        super(msg);
+    }
+
+}
