--- src/main/java/com/souljit2/selleverything/post/PostCategory.java
@@ -1,11 +1,0 @@
-package com.souljit2.selleverything.post;
-
-public enum PostCategory {
-    CLOTHES,
-    INSTRUMENTS,
-    VEHICLES,
-    PHONES,
-    ACCESSORIES,
-    PROPS,
-    ETCS
-}
