--- src/main/java/com/souljit2/selleverything/service/AuthService.java
@@ -1,14 +1,0 @@
-package com.souljit2.selleverything.service;
-
-import com.souljit2.selleverything.model.MemberDTO;
-import com.souljit2.selleverything.model.SignInRequestDTO;
-
-public interface AuthService {
-
-    public void signUp(MemberDTO newMemberInfo);
-
-    public void signIn(SignInRequestDTO signInInfo);
-
-    public int getRequestMemberId();
-
-}
