--- src/main/java/com/souljit2/selleverything/service/PostServiceImpl.java
@@ -1,65 +1,0 @@
-package com.souljit2.selleverything.service;
-
-import com.souljit2.selleverything.constants.CacheNames;
-import com.souljit2.selleverything.constants.DomainNames;
-import com.souljit2.selleverything.mapper.PostMapper;
-import com.souljit2.selleverything.model.PostDTO;
-import lombok.AllArgsConstructor;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
-
-@Service
-@AllArgsConstructor
-public class PostServiceImpl implements PostService {
-
-    PostMapper postMapper;
-
-    AuthService authService;
-
-    SessionAuthServiceImpl sessionAuthService;
-
-    @Override
-    public PostDTO getPostById(int id) {
-        return postMapper.getPostById(id);
-    }
-
-    @Cacheable(value = CacheNames.POST)
-    @Override
-    public List<PostDTO> getPostsByQueryString(Map<String, String> queryMap) {
-        return postMapper.getPosts(queryMap);
-    }
-
-    @CacheEvict(value = CacheNames.POST, allEntries = true)
-    @Override
-    public void createPost(PostDTO newPost) {
-        int memberIdBySession = authService.getRequestMemberId();
-        postMapper.createPost(PostDTO.builder()
-            .postTitle(newPost.getPostTitle())
-            .postContents(newPost.getPostContents())
-            .postItemName(newPost.getPostItemName())
-            .postItemPrice(newPost.getPostItemPrice())
-            .postCategory(newPost.getPostCategory())
-            .memberIdFk(memberIdBySession)
-            .build()
-        );
-    }
-
-    @CacheEvict(value = CacheNames.POST, allEntries = true)
-    @Override
-    public void updatePostById(int id, PostDTO newPost) {
-        int sessionMemberId = sessionAuthService.getRequestMemberId();
-        postMapper.updatePostById(id, sessionMemberId, newPost);
-    }
-
-    @CacheEvict(value = CacheNames.POST, allEntries = true)
-    @Override
-    public void deletePostById(int id) {
-        int sessionMemberId = sessionAuthService.getRequestMemberId();
-        postMapper.deletePostById(id, sessionMemberId);
-    }
-}
